improved for c++98

+++ c06/ex01/Serializer.hpp
@@ -1,7 +1,6 @@
 #ifndef SERIALIZER_HPP
 #define SERIALIZER_HPP
 
-#include <cstdint>
 #include "Data.hpp"
 
 class Serializer {
@@ -12,8 +11,8 @@ class Serializer {
         ~Serializer();
 
     public:
-        static uintptr_t serialize(Data* ptr);
-        static Data* deserialize(uintptr_t raw);
+        static unsigned long serialize(Data* ptr);
+        static Data* deserialize(unsigned long raw);
 };
 
 #endif
